requirements.txt
aiohttp==3.11.12
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
frozenlist==1.5.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
multidict==6.1.0
protobuf==5.29.5
temporalio==1.20.0
types-protobuf==5.26.0.20240422
typing_extensions>=4.12.2
Werkzeug==3.0.6
yarl==1.18.3
pydantic-ai>=1.61.0
python-dotenv>=1.0.1
logfire>=4.24.0
